--- client-copy/src/Home.js
@@ -1,23 +1,0 @@
-import image from "./images/mixed.png";
-import "./landing_page.css";
-
-function Home() {
-  return (
-    <main>
-      <h1>Spice Guys 🌿</h1>
-      <div>
-        <img src={image} alt="spices" />
-        <div>
-          <h2>Spice Guys is a spice shop that catalogs a variety of spices and herbs.</h2>
-          <div>
-         <button>hello</button>
-         
-         </div>
-        </div>
-      </div>
-    </main>
-  );
-}
-
-
-export default Home;
