+++ client-copy/src/components/screens/AuthScreens.css
@@ -1,0 +1,104 @@
+.login {
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: black;
+}
+
+.login-form {
+    width: 380px;
+    padding: 1.5rem;
+    box-shadow: 0 1rem 2rem rgba(0, 0, 0, 0.2);
+    background: #fff;
+}
+
+.login-title {
+    text-align: center;
+    margin-bottom: 1rem;
+}
+
+.login-forgotpassword {
+    font-size: 0.7rem;
+}
+
+.login-subtext {
+    font-size: 0.7rem;
+    display: block;
+    margin: 0.5rem 0;
+}
+
+.register {
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: black;
+}
+
+.register-form {
+    width: 380px;
+    padding: 1.5rem;
+    box-shadow: 0 1rem 2rem rgba(0, 0, 0, 0.2);
+    background: #fff;
+}
+
+.register-title {
+    text-align: center;
+    margin-bottom: 1rem;
+}
+
+.register-subtext {
+    font-size: 0.7rem;
+    display: block;
+    margin: 0.5rem 0;
+}
+
+.forgotpassword {
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.forgotpassword-form {
+    width: 380px;
+    padding: 1.5rem;
+    box-shadow: 0 1rem 2rem rgba(0, 0, 0, 0.2);
+    background: #fff;
+}
+
+.forgotpassword-title {
+    text-align: center;
+    margin-bottom: 1rem;
+}
+
+.forgotpassword-subtext {
+    font-size: 0.7rem;
+    display: block;
+    margin: 0.5rem 0;
+}
+
+
+.resetpassword {
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.resetpassword-form {
+    width: 380px;
+    padding: 1.5rem;
+    box-shadow: 0 1rem 2rem rgba(0, 0, 0, 0.2);
+    background: #fff;
+}
+
+.resetpassword-title {
+    text-align: center;
+    margin-bottom: 1rem;
+}
