+++ client-copy/src/components/screens/LandingPage.css
@@ -1,5 +1,3 @@
-/* this is the .css file that styles the Home.js folder */
-
 main {
     background: #000;
     font-family: 'Roboto', sans-serif;
