--- client-copy/src/components/screens/Login.css
@@ -1,30 +1,0 @@
-.login {
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: black;
-}
-
-.login-form {
-    width: 380px;
-    padding: 1.5rem;
-    box-shadow: 0 1rem 2rem rgba(0, 0, 0, 0.2);
-    background: #fff;
-}
-
-.login-title {
-    text-align: center;
-    margin-bottom: 1rem;
-}
-
-.login-forgotpassword {
-    font-size: 0.7rem;
-}
-
-.login-subtext {
-    font-size: 0.7rem;
-    display: block;
-    margin: 0.5rem 0;
-}
