+++ client-copy/src/components/screens/Login.js
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import axios from "axios";
 import { Link } from "react-router-dom";
-import "./Login.css";
+import "./AuthScreens.css";
 
 const Login = ({ history }) => {
