--- client-copy/src/components/screens/Register.css
@@ -1,26 +1,0 @@
-.register {
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: black;
-}
-
-.register-form {
-    width: 380px;
-    padding: 1.5rem;
-    box-shadow: 0 1rem 2rem rgba(0, 0, 0, 0.2);
-    background: #fff;
-}
-
-.register-title {
-    text-align: center;
-    margin-bottom: 1rem;
-}
-
-.register-subtext {
-    font-size: 0.7rem;
-    display: block;
-    margin: 0.5rem 0;
-}
