--- client-copy/src/components/screens/ResetPassword.css
@@ -1,19 +1,0 @@
-.resetpassword {
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.resetpassword-form {
-    width: 380px;
-    padding: 1.5rem;
-    box-shadow: 0 1rem 2rem rgba(0, 0, 0, 0.2);
-    background: #fff;
-}
-
-.resetpassword-title {
-    text-align: center;
-    margin-bottom: 1rem;
-}
