+++ client-copy/src/components/screens/ResetPassword.js
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Link } from "react-router-dom";
 import axios from "axios";
-import "./ResetPassword.css";
+import "./AuthScreens.css";
 
 const ResetPassword = ({ match }) => {
     const [password, setPassword] = useState("");
