+++ client-copy/src/components/screens/SpicePage.js
@@ -1,0 +1,23 @@
+// import React from 'react'
+// import Card from "react-bootstrap/Card";
+// import 'bootstrap/dist/css/bootstrap.min.css';
+
+// function ShowSpice() {
+
+//     function renderCard(card, index) {
+//         return (
+//             <Card className="bg-dark text-dark">
+//                 <Card.Img src="https://place-puppy.com/200x200" alt="Card image" />
+//                 <Card.ImgOverlay>
+//                     <Card.Title>InstaClone!  Cloned Food!  Delivered!</Card.Title>
+//                     <Card.Text>
+//                         Dogs are not food just cute.
+//                     </Card.Text>
+//                 </Card.ImgOverlay>
+//             </Card>
+//         )
+//     }
+//     return <div className='grid'>{cardInfo.map(renderCard)}</div>
+// }
+
+// export default ShowSpice
