--- client-copy/src/landing_page.css
@@ -1,89 +1,0 @@
-/* this is the .css file that styles the Home.js folder */
-
-main {
-    background: #000;
-    font-family: 'Roboto', sans-serif;
-    font-size: 1em;
-    font-weight: 300;
-    color: #fff;
-    text-align: center;
-    margin: 0;
-    padding: 0;
-    height: 100vh;
-}
-
-
-h1 {
-    font-family: 'Roboto', sans-serif;
-    font-size: 3em;
-    font-weight: 300;
-    color: #fff;
-    text-align: center;
-    margin: 0;
-    padding: 0;
-}
-
-h2 {
-    font-family: 'Roboto', sans-serif;
-    font-size: 2em;
-    font-weight: 300;
-    color: #fff;
-    text-align: center;
-    margin: 0;
-    padding: 2px 0 0 0;
-}
-
-button {
-    background-color: #4CAF50; /* Green */
-    border: none;
-    color: white;
-    padding: 15px 32px;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 16px;
-  }
-
-  
-
-  /* Style the navigation menu */
-.topnav {
-    overflow: hidden;
-    background-color: #333;
-    position: relative;
-  }
-  
-  /* Hide the links inside the navigation menu (except for logo/home) */
-  .topnav #myLinks {
-    display: none;
-  }
-  
-  /* Style navigation menu links */
-  .topnav a {
-    color: white;
-    padding: 14px 16px;
-    text-decoration: none;
-    font-size: 17px;
-    display: block;
-  }
-  
-  /* Style the hamburger menu */
-  .topnav a.icon {
-    background: black;
-    display: block;
-    position: absolute;
-    right: 0;
-    top: 0;
-  }
-  
-  /* Add a grey background color on mouse-over */
-  .topnav a:hover {
-    background-color: #ddd;
-    color: black;
-  }
-  
-  /* Style the active link (or home/logo) */
-  .active {
-    background-color: #04AA6D;
-    color: white;
-  }
